[refactor] Changed names, deleted the previous ai controller files and updated routes (#74)
