+++ backend/src/routes/results.routes.ts
@@ -1,8 +1,7 @@
 import { Router } from "express";
 import multer from "multer";
-import * as screening from "../controllers/screening.controller.js";
 import * as results from "../controllers/results.controller.js";
-import * as ai from "../controllers/ai.controller.js"
+import * as ai from "../controllers/screening.controller.js"
 
 const router = Router();
 const upload = multer({
@@ -12,12 +11,11 @@ const upload = multer({
 
 router.get("/", results.list);
 router.post("/", results.create);
-router.post("/screenings/run", upload.single("jobDescriptionFile"), screening.runScreening);
+router.post("/screenings/run", upload.single("jobDescriptionFile"), ai.runScreeningController);
 router.post("/screenings", results.createScreeningRun);
 router.get("/history", results.getScreeningHistory);
 router.get("/job_posts/:jobPostId", results.getScreeningByJobPostId);
 router.get("/:jobPostId/:candidateId", results.getById);
 router.delete("/:jobPostId/:candidateId", results.deleteById);
 
-router.post("/test", ai.runScreeningController);
 export default router;
