+++ backend/src/types/ai.types.ts
